No nils, use symbols instead
nil have no meaning, symbol is meaningfull. This change will allow to
check if published event is the first one in given stream (expect version :none).
The nil value that was used to mark we do not care about what is currently in stream
has been replaced with symbol :any.