Release management replacement for monorepo.
Previously we've used `rake release` on each of the gems that handled:
- setting git tag
- building gem pkg
- pushing pkg to rubygems

That was automated with Bundler and
https://github.com/bundler/bundler/blob/46b4b2f23c3d50314bed694fc634380798ba0721/lib/bundler/gem_helper.rb.

We've also had to bump gem version and update dependencies (if neeeded).

Now that we have monorepo we don't have this out of the box. This
commit brings back release automation with `make release`:
- version is read from RES_VERSION and set in all gems and gemspecs
  (dependencies version)
- git tag is set and pushed
- pkg build and rubygems push executed for each gem