Unique index on events in streams
Our codebase relies on non-duplicated events. Right now it is not
possible to publish same event twice. But if we add append_to_stream, it
will be possible to have same event in many streams.

Don't mutate #detect_pkey_index_violated because it has logic which work
on different databases and we can't run mutant across many DBs at the
same time easily (we could have many connections but not worth it imho).