Make it possible to subscribe any object
As long as the dispatcher allows and knows how to handle it.
Remove passing proxy_for around and the whole logistics.

Right now I used inheritance but perhaps dispatchers should be a
pipeline/middlewares pattern in the future. If one dispatcher does not
know how to handle dispatching to a certain subscriber maybe next in
line knows and if not maybe another one would know until the last one.

I am not convinced static verification during a subscription is useful.
I would consider removing #verify as well. I kept it because we had
basic verifications so far.

Additional benefit of that approach is that we have the original subscriber in
the list of subscribers and we could get their names etc and display in
a rake task.