Events still not persisted if nested transaction used
A developer can start a transaction outside of RES. We need to handle
this situation properly!

http://api.rubyonrails.org/classes/ActiveRecord/Transactions/ClassMethods.html#module-ActiveRecord::Transactions::ClassMethods-label-Nested+transactions

transaction calls can be nested. By default, this makes all database statements in the nested transaction block become part of the parent transaction. For example, the following behavior may be surprising:

User.transaction do
  User.create(username: 'Kotori')
  User.transaction do
    User.create(username: 'Nemu')
    raise ActiveRecord::Rollback
  end
end

creates both “Kotori” and “Nemu”. Reason is the ActiveRecord::Rollback
exception in the nested block does not issue a ROLLBACK. Since these
exceptions are captured in transaction blocks, the parent block does not see
it and the real transaction is committed.

AAAAAAAAAAAAAaaaaaaaa !!!!

In order to get a ROLLBACK for the nested transaction you may ask for a real
sub-transaction by passing requires_new: true. If anything goes wrong, the
database rolls back to the beginning of the sub-transaction without rolling
back the parent transaction. If we add it to the previous example:

User.transaction do
  User.create(username: 'Kotori')
  User.transaction(requires_new: true) do
    User.create(username: 'Nemu')
    raise ActiveRecord::Rollback
  end
end

only “Kotori” is created. This works on MySQL and PostgreSQL.
SQLite3 version >= '3.6.8' also supports it.