Make sure it works without Sprockets.
Sprockets/Asset Pipeline/Webpacker are not needed since we already have
existing build setup for Elm. For release we only need to serve compiled
JS bundle. That is just enough for plain static assets middleware.

Since we merge RES::Browser assets with the application assets this
engine is mounted on, a more "unique" file name is chosen than
bundle.js.