alternative implementation of normalize_to_array
it's mostly to make mutant more happy and a
work-around for the mutant bug:
https://github.com/mbj/mutant/issues/724

I hope this implementation does exactly the same as
the previous, that was my assumption, but please
correct/revert if I'm wrong.

It might be a subjective opinion, but I think this
code exposes the problem of our API here. We can't
really be sure until this moment if we have 1 event
or a list of events.

This leads to naming problems,is it event or events?

I'm not sure what's the solution and if anyone else
sees this as a problem.

option 1: we force people to use [event] as an input.
option 2: we have two new methods:
link_event_to_stream
link_events_to_stream
(here we deprecate the current one)

In my opinion any of those would be better than the
existing solution. However both would be breaking
the public API :(