normalize events arguments as early as possible
this is to avoid being unsure later in the code what
objects do we deal with

as an addition to my previous commit message:
if we insist on having such public API (popular in
the Rails  - allowing a single object or array at the same time) then maybe let's have it at the
RailsEventStore level, but immediatelly later at
RubyEventStore make it explicit and do the
normalization in there.

This would be in the spirit of having RubyEventStore
free of any Rails magic or popular in the community
Ruby tricks - they could still live at the
RailsEventStore level.

I do respect the fact, that not everyone has to
agree with my opinion that such API is not perfect.

My opinion here is similar for allowing passing
event_ids vs event objects sometimes, if we do so.
I don't think it's a good idea overall.