Move TestEvent to spec_helper
We'll need it in two spec files - client_spec & new_client_spec.

(cherry pick commit b1713fa6)

We've extracted client spec to shared_examples in fluent-api branch, and
we're planning to move specs from spec/actions dir as well. It makes a
lot of sense to make this changes on master to avoid potential
conflicts.