Explicit DESC query test
Sometimes the layout of rows in database is identical to "id ASC" or
"id DESC". In such case mutation (of `order('id ASC')` to nothing
can survive because rows will be received accidentally in the right
order. There is no good way to kill such mutation. We check that our
query explicitly states DESC/ASC order to prevent it and to double check
that we know what we are doing. Perhaps it would not fail if we had 1000
objects stored and randoms would be less likely?