Mutant questioning query choice.
I was wondering why we specifically read backward when looking for match
of expectation in a stream and the only things I recall is that:
- read_events_backward/forward had a default limit of 100 events and...
- in tests where 100+ events were published, given the most interesting
  ones in SUT were the last (after setup), you'd get false negative (if
  your event was at least 101-st)
- which is no longer a case in unbounded read here

Unbounded read screams "scary" but I think in test context that still
should be <1000 events which is okayish. Also using primarily InMemoryRepository
in testing that would mean we've already loaded all those events in
memory (:shrug:).