Use inline adapter when async adapter not available
in Rails 4.2.*

It doesn't really test much (which was revealed by mutation testing and
thus changed to :async) because it executes the job in the same thread
and thus getting the metadata that we prepared for sync-handlers. So not
really testing what we have in mind.

We should include additional integration tests which would verify with
a few other adapters for safety.

Issue: #346