Not a development-only dependency :see_no_evil:
We used to think about it as a non-event-store dependent gem but this
has changed with the introduction of more sophisticated RES apis (like
batching).

In a way, we depend on RES here. You can still substitute RES for
something else of identical read API. Although that might be not
practical.