Ensure Sidekiq::Worker works with AsyncHandler helper.
When serializing payload for background job system with JSON,
information about symbols as keys is lost.

ActiveJob persists the information about symbols in payload. This
information is used to convert those keys back to symbols when the
payload is passed back to worker.

When using plain Sidekiq::Worker without ActiveJob wrapper, this
mechanism is no longer in place. The polyfill for such behaviour is
rather simple in case of SerializedRecord thus we provide it for
convenience in RailsEventStore::AsyncHandler.

When you don't use RailsEventStore::AsyncHandler, the responsibility of
adjusting received payload for event_store.deserialize() is on you.

Closes #507.