NullMapper is essentially a default mapper with no serialization.
Loosen SerializedRecord data/metadata type constraint to allow null
serialization: in-memory repository would not require such constraint
and that resposibility can be moved to serializer (although no need
to do so for upstream JSON and YAML)

This change makes NullMapper do the conversion to and from
SerializedRecord: pro is that we can expect to_h, hash and event_type
to exhibit correct behaviour (as compared to output of a non-null mapper).

The con is that it makes additional allocations when converting from
events to and from records. Probably unavoidable given how
SerializedRecord diverged from an Event.