Move cleaning after releasing a lock
I'm not 100% sure whether it's a good move, because if we clean the
records after releasing, then we remove the records from the split
some other process may be working on. But there should be no technical
problems with it, as far as I know. The argument to release early is
that cleanup may take some time, and we don't want to not be
processing the records during that cleaning.