Failing test case to illustate the problem
Broker is broken with regard to custom event type. It does not have a
resolver at hand and at the same time wotks with events and records.

It should probably continue passing event_type to Subscriptions. Instead
event type resolver is to be lifted up in the stack, closer to the
Client facade.