Ignore known-failing mutations
These will unnecessarily hang on max timeout -- not sure why such
behaviour is present in mutant yet.

Without timeout-as-kill these clearly fail (but not on timeout).

Failing is related to non-sqlite-specific code (we mutate on one
database only) and to ROM-specific setup (these failures are weird,
external library so no much interest to improve here).