Introduce RubyEventStore::Serializers::YAML as default serializer
Up until now regular `YAML` module shipped with Ruby was used as a
default serializer. However, with release of Ruby 3.1 this module
started to use Psych 4 instead of Psych 3. This, in turn, resulted in a
breaking change when non-primitive values were serialized and
deswrializes, as since version 4 Psych does "safe loading" by default,
which disallows deserializing objects such as BigDecimal or Time.

As this breaking change also leaks into RES (old events would no longer
be correctly deserialized afer upgrading to Ruby 3.1), a proxy
serializer is introduced. It checks if `unsafe_load` method of `YAML` is
present (which indicates Psych 4) and uses it if available, falling back
to plain old `YAML.load` if it's not present.