Either test the defaults or ignore them
Mutant likes to remind that setting defaults is ineffective, when we
override them in tests. This points out false sense of security: we
cannot rely on the code we didn't test and have released.

Resolution is to test with initializer default. If we had another
serializer with specified behaviour, then it would make sense to pass it
explicitly in tests.