Ensure version in broadcasted to frontend
While mutant suggests we could use Browser::VERSION, which is in sync
with RubyEventStore::VERSION, I'd rather keep it explicit in case it
diverges in the future. Thus ignoring but adding unit.

Alternative would be to re-create some unrealistic scenario (diverging
RES and Browser versions) or checking implementation details. Which all
seem a bit too much.