Add Tempo dynamically controlling the batch size
Rationale: The static batch size (let's say 100) is a problem when we
sometimes have very very big messages. We fetch 100 messages, few are
big ones, we crash (OOM). We start again, we fetch 100 messages, we
crash again. Whereas we could (just after starting the outbox process
fetch only one message, assuming that this could be the "very big one"
because of which we crashed.