APP_TEMPLATE provides well–suited client and db
Until now, creating app with -m https://rails_event_store.org/new
template created initializer with default client and default database,
no matter the database driver used. Meaning:
  * RailsEventStore::Client.new with default YAML event serializer for
    EventRepository
  * binary columns for keeping event's data and metadata

Based on our experience, we know that for app running postgresql, the
best idea is to use the jsonb column type. However, RES client
configuration must follow that and we had observed that this piece can be
tricky for many users. Because of that, we have introduced JSONClient
which inherits from Client, but uses correct serializer configuration,
accomodating column. For the time being it's RubyEventStore::NULL. We
already now that serializer for use with json(b) columns will change with
RES 3.0.0, but we expect that this change will be transparent for new and
easy to introduce for existing users.