Don't suggest use of NULL serializer
It's better to rely on explicit serializer for data and metadata.
So that we pass string values in a flat hash to sidekiq.
Thus we suggest RubyEventStore::Serializers::YAML or simply JSON as
a serializer, so 3rd party (sidekiq) won't decide on how to handle
serialization/deserialization.

There's an added benefit of doing so: serialization/deserialization
would occur just once when using the same serializer for both scheduler
and repository.