Drop testing RES release on Truffleruby
Truffleruby workflows became very unstable recently. They randomly fail
too frequently, without any or with very little changes in the RES core.

Truffleruby is rather a nice-to-have than must-have for RES.
There were a few reproducible issues in the past that were solved in
Truffleruby repository thanks to our testing pipeline.

I'm not aware of any production RES deployment running on Truffle at the
moment. I don't thnk any of the maintainers have bandwidth to maintain
and resolve current issues with Truffle implementation, which are more
difficult to identify and reproduce. I would prefer to spend my limited
time woking on RES instead.

Perhaps the solution is to separate Truffleruby into separate test
workflow, which does not affect statuses of submitted PRs and commmits
and runs side-by side as a separate project. However it would be nice
to have a justification for this first.